+++ flutter_navigation_generator/lib/src/models/route_config.dart
@@ -8,7 +8,9 @@ class RouteConfig {
   final bool generatePageRoute;
   final bool generateMethod;
   final bool routeNameIsDefinedByAnnotation;
+  final bool methodNameIsDefinedByAnnotation;
   final String routeName;
+  final String methodName;
   final String constructorName;
   final ImportableType routeWidget;
   final ImportableType? returnType;
@@ -17,15 +19,19 @@ class RouteConfig {
   final List<ImportableType> parameters;
   final Map<String, dynamic> defaultValues;
 
+  bool get routeNameContainsParameters => routeName.contains(':');
+
   RouteConfig({
     required this.routeWidget,
     this.routeName = '',
+    this.methodName = '',
     this.constructorName = '',
     this.pageType,
     this.isFullscreenDialog = false,
     this.generatePageRoute = true,
     this.generateMethod = true,
     this.routeNameIsDefinedByAnnotation = false,
+    this.methodNameIsDefinedByAnnotation = false,
     this.returnType,
     this.navigationType = NavigationType.push,
     this.parameters = const [],
@@ -39,6 +45,9 @@ class RouteConfig {
       'generateMethod': generateMethod,
       'constructorName': constructorName,
       'routeName': routeName,
+      'methodName': methodName,
+      'routeNameIsDefinedByAnnotation': routeNameIsDefinedByAnnotation,
+      'methodNameIsDefinedByAnnotation': methodNameIsDefinedByAnnotation,
       'returnType': returnType?.toMap(),
       'pageType': pageType?.toMap(),
       'routeWidget': routeWidget.toMap(),
@@ -55,24 +64,19 @@ class RouteConfig {
       generateMethod: map['generateMethod'] ?? false,
       constructorName: map['constructorName'] ?? '',
       routeName: map['routeName'] ?? '',
+      methodName: map['methodName'] ?? '',
+      routeNameIsDefinedByAnnotation: map['routeNameIsDefinedByAnnotation'] ?? '',
+      methodNameIsDefinedByAnnotation: map['methodNameIsDefinedByAnnotation'] ?? '',
       routeWidget: ImportableType.fromMap(map['routeWidget']),
-      returnType: map['returnType'] != null
-          ? ImportableType.fromMap(map['returnType'])
-          : null,
-      pageType: map['pageType'] != null
-          ? ImportableType.fromMap(map['pageType'])
-          : null,
+      returnType: map['returnType'] != null ? ImportableType.fromMap(map['returnType']) : null,
+      pageType: map['pageType'] != null ? ImportableType.fromMap(map['pageType']) : null,
       navigationType: NavigationType.values[map['navigationType']],
-      parameters: List<ImportableType>.from(map['parameters']?.map(
-              (dynamic x) => ImportableType.fromMap(x as Map<String, dynamic>))
-          as Iterable),
-      defaultValues:
-          map['defaultValues'] as Map<String, dynamic>? ?? <String, dynamic>{},
+      parameters: List<ImportableType>.from(map['parameters']?.map((dynamic x) => ImportableType.fromMap(x as Map<String, dynamic>)) as Iterable),
+      defaultValues: map['defaultValues'] as Map<String, dynamic>? ?? <String, dynamic>{},
     );
   }
 
   String toJson() => json.encode(toMap());
 
-  factory RouteConfig.fromJson(String source) =>
-      RouteConfig.fromMap(json.decode(source));
+  factory RouteConfig.fromJson(String source) => RouteConfig.fromMap(json.decode(source));
 }