+++ flutter_navigation_generator/lib/src/utils/case_utils.dart
@@ -1,6 +1,6 @@
 class CaseUtil {
   final RegExp _upperAlphaRegex = RegExp(r'[A-Z]');
-  final RegExp _symbolRegex = RegExp(r'[ ./_\-]');
+  final RegExp _symbolRegex = RegExp(r'[ ./_\-:]');
 
   late String originalText;
   late List<String> _words;
@@ -22,9 +22,7 @@ class CaseUtil {
 
     for (var i = 0; i < text.length; i++) {
       final char = String.fromCharCode(text.codeUnitAt(i));
-      final nextChar = i + 1 == text.length
-          ? null
-          : String.fromCharCode(text.codeUnitAt(i + 1));
+      final nextChar = i + 1 == text.length ? null : String.fromCharCode(text.codeUnitAt(i + 1));
 
       if (_symbolRegex.hasMatch(char)) {
         continue;
@@ -32,9 +30,7 @@ class CaseUtil {
 
       sb.write(char);
 
-      final isEndOfWord = nextChar == null ||
-          (_upperAlphaRegex.hasMatch(nextChar) && !isAllCaps) ||
-          _symbolRegex.hasMatch(nextChar);
+      final isEndOfWord = nextChar == null || (_upperAlphaRegex.hasMatch(nextChar) && !isAllCaps) || _symbolRegex.hasMatch(nextChar);
 
       if (isEndOfWord) {
         words.add(sb.toString());
@@ -55,8 +51,7 @@ class CaseUtil {
 
   String get upperCamelCase => _uppserCamelCase();
 
-  String _getSnakeCase({String separator = '_'}) =>
-      _words.map((word) => word.toLowerCase()).toList().join(separator);
+  String _getSnakeCase({String separator = '_'}) => _words.map((word) => word.toLowerCase()).toList().join(separator);
 
   String _getCamelCase({String separator = ''}) {
     final words = _words.map(_upperCaseFirstLetter).toList();
@@ -71,7 +66,7 @@ class CaseUtil {
   }
 
   String _getTextWithoutSuffix() {
-    final delimiters = [' ', '-', '_', '.', '/'];
+    final delimiters = [' ', '-', '_', '.', '/', ':'];
     if (delimiters.any(originalText.endsWith)) {
       return originalText.substring(0, originalText.length - 1);
     } else {